requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
boto3==1.14.12
botocore==1.17.12
click==7.1.2
colorama==0.4.3
Django==3.0.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-social-share==1.4.0
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
Pillow==7.1.2
pycodestyle==2.6.0
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
